Modifications and Adding IA Fetch
Modified the order of project items in the grid. Added my first IA document in the CMS and noticed it was only returning an ID from the API, so added the fetchMedia action creator and reducer to handle grabbing the URL. Tweaked the image-based items to use "url" as well in state.