Adding Major CTAs to Home + Media Link Update
Decided to add the major CTAs (e.g. View Project, View Source, View Planning) to the items on the home screen for now. I can break it all out into individual pages at a later date.

Also discovered the URL I was using for media docs was incorrect and changed that to the source url.